+++ projects/ngx-progressbar/src/lib/ng-progress-ref.ts
@@ -3,31 +3,48 @@ import {
   inject,
   signal,
   effect,
+  output,
   computed,
   untracked,
+  numberAttribute,
+  input,
   Signal,
+  InputSignal,
   WritableSignal,
-  EffectCleanupRegisterFn
+  OutputEmitterRef,
+  EffectCleanupRegisterFn,
+  InputSignalWithTransform
 } from '@angular/core';
+import { outputToObservable } from '@angular/core/rxjs-interop';
 import {
   Observable,
-  Subject,
   Subscription,
   BehaviorSubject,
   of,
   tap,
   delay,
   timer,
-  switchMap,
+  filter,
   finalize,
+  switchMap,
   takeUntil,
   EMPTY
 } from 'rxjs';
-import { NgProgressOptions, NG_PROGRESS_OPTIONS } from './ng-progress.model';
+import { NgProgressOptions, NG_PROGRESS_OPTIONS, NgProgressRefOptions } from './ng-progress.model';
 
 enum TriggerType {
   START = 'START',
   COMPLETE = 'COMPLETE'
+}
+
+function minAttribute(value: string | number): number {
+  const min: number = numberAttribute(value) || 0;
+  return (min < 100 && min >= 0) ? min : 0;
+}
+
+function maxAttribute(value: string | number): number {
+  const max: number = numberAttribute(value) || 100;
+  return (max > 0 && max <= 100) ? max : 100;
 }
 
 @Directive({
@@ -37,7 +54,27 @@ enum TriggerType {
 })
 export class NgProgressRef {
 
-  private readonly _config: WritableSignal<NgProgressOptions> = signal<NgProgressOptions>(inject(NG_PROGRESS_OPTIONS));
+  private readonly defaultOptions: NgProgressOptions = inject(NG_PROGRESS_OPTIONS);
+
+  min: InputSignalWithTransform<number, number | string> = input<number, number | string>(this.defaultOptions.min, { transform: minAttribute });
+  max: InputSignalWithTransform<number, number | string> = input<number, number | string>(this.defaultOptions.max, { transform: maxAttribute });
+  speed: InputSignalWithTransform<number, number | string> = input<number, number | string>(this.defaultOptions.speed, { transform: numberAttribute });
+  trickleSpeed: InputSignalWithTransform<number, number | string> = input<number, number | string>(this.defaultOptions.trickleSpeed, { transform: numberAttribute });
+  fadeOutSpeed: InputSignalWithTransform<number, number | string> = input<number, number | string>(this.defaultOptions.fadeOutSpeed, { transform: numberAttribute });
+  debounceTime: InputSignalWithTransform<number, number | string> = input<number, number | string>(this.defaultOptions.debounceTime, { transform: numberAttribute });
+  trickleFunc: InputSignal<(n: number) => number> = input(this.defaultOptions.trickleFunc);
+
+  config: Signal<NgProgressRefOptions> = computed(() => {
+    return {
+      max: this.max(),
+      min: this.min(),
+      speed: this.speed(),
+      trickleSpeed: this.trickleSpeed(),
+      fadeOutSpeed: this.fadeOutSpeed(),
+      trickleFunc: this.trickleFunc(),
+      debounceTime: this.debounceTime()
+    }
+  });
 
   private _progress: WritableSignal<number> = signal<number>(0);
 
@@ -47,17 +84,13 @@ export class NgProgressRef {
 
   progress: Signal<number> = computed(() => this._progress());
 
-  config: Signal<NgProgressOptions> = computed(() => this._config());
-
-  private _trigger: BehaviorSubject<TriggerType> = new BehaviorSubject<TriggerType>(null);
-
   // Progress start source event (used to cancel onComplete delays)
-  private readonly _started: Subject<void> = new Subject<void>();
-  readonly started: Observable<void> = this._started.asObservable();
+  started: OutputEmitterRef<void> = output<void>();
 
   // Progress ended source event
-  private readonly _completed: Subject<void> = new Subject<void>();
-  readonly completed: Observable<void> = this._completed.asObservable();
+  completed: OutputEmitterRef<void> = output<void>();
+
+  private _trigger: BehaviorSubject<TriggerType> = new BehaviorSubject<TriggerType>(null);
 
   constructor() {
     let sub$: Subscription;
@@ -65,23 +98,20 @@ export class NgProgressRef {
       const config: NgProgressOptions = this.config();
 
       untracked(() => {
-        sub$?.unsubscribe();
-
         sub$ = this._trigger.pipe(
+          filter((trigger: TriggerType) => !!trigger),
           switchMap((trigger: TriggerType) => {
             if (trigger === TriggerType.START) {
               return timer(config.debounceTime).pipe(
                 switchMap(() => this.onTrickling(config))
               );
-            } else if (trigger === TriggerType.COMPLETE) {
-              return this.onComplete(config);
             }
-            return EMPTY;
+            return this.onComplete(config);
           })
         ).subscribe();
-      })
 
-      onCleanup(() => sub$?.unsubscribe());
+        onCleanup(() => sub$?.unsubscribe());
+      });
     });
   }
 
@@ -89,7 +119,7 @@ export class NgProgressRef {
    * Start the progress
    */
   start(): void {
-    this._started.next();
+    this.started.emit();
     this._trigger.next(TriggerType.START);
     this._active.set(true);
   }
@@ -125,13 +155,6 @@ export class NgProgressRef {
   }
 
   /**
-   * Set config
-   */
-  setConfig(config: NgProgressOptions): void {
-    this._config.set({ ...this.config(), ...config });
-  }
-
-  /**
    * Clamps a value to be between min and max
    */
   private clamp(n: number): number {
@@ -152,27 +175,27 @@ export class NgProgressRef {
    * Completes then resets the progress
    */
   private onComplete(config: NgProgressOptions): Observable<unknown> {
+    // If it's not active no need to complete
     if (!this.active()) {
       return EMPTY;
     }
 
     // Emit completed
-    this._completed.next();
+    this.completed.emit();
 
     return of({}).pipe(
       // Complete the progress
       tap(() => this._progress.set(100)),
-
       // Deactivate the progress after a tiny delay
-      delay(config.speed * 1.7),
+      delay(config.speed + 140),
       tap(() => this._active.set(false)),
 
       // Use a tiny delay before resetting
-      delay(config.speed),
+      delay(config.fadeOutSpeed),
       // Force the progress to reset even it got cancelled
       finalize(() => this._progress.set(0)),
       // Cancel any of the finalizing delays if the progress has started again
-      takeUntil(this._started)
+      takeUntil(outputToObservable(this.started))
     );
   }
 }