+++ projects/ngx-progressbar/src/lib/ng-progress.model.ts
@@ -1,18 +1,22 @@
 import { InjectionToken, Provider } from '@angular/core';
 import { defaultOptions } from './ng-progress-default';
 
-export interface NgProgressOptions {
+export interface NgProgressRefOptions {
+  max?: number;
+  min?: number;
+  speed?: number;
+  trickleSpeed?: number;
+  fadeOutSpeed?: number;
+  trickleFunc?: (n: number) => number;
+  debounceTime?: number;
+}
+
+export type NgProgressOptions = NgProgressRefOptions & {
   spinnerPosition?: 'left' | 'right';
   direction?: 'ltr+' | 'ltr-' | 'rtl+' | 'rtl-';
   relative?: boolean;
   flat?: boolean;
   spinner?: boolean;
-  max?: number;
-  min?: number;
-  speed?: number;
-  trickleSpeed?: number;
-  trickleFunc?: (n: number) => number;
-  debounceTime?: number;
 }
 
 export const NG_PROGRESS_OPTIONS: InjectionToken<NgProgressOptions> = new InjectionToken<NgProgressOptions>('NG_PROGRESS_OPTIONS', {